fix(mediawiki): Fix field type of MediaWikiUser.user_name
The field in the database is a VARBINARY.  Using pewee's `TextField` on
that returns values as `"bytearray(b'Name')" (i.e. as a Python `str`
containing the `repr` of the `binary` field data, which is just wild).

Changing the field type to `BlobField` results in `MediaWikiUser.user_name`
being a `binary`-encoded string, as expected.