refactor(connectdb): Use pymysql instead of mysql.connector
Originally, `chimedb.core` used the C-extension `MySQLdb` to
connect to the database.  Eventually we got fed up with having to
compile against the mysql client CAPI library, so we switched to
`mysql.connector`, which is a pure-python alternative provided by
Oracle (owners of MySQL).

However, that was all back in the Python2 days.  Now there's `pymysql`
which is `peewee`'s preferred way of interfacing with MySQL.  So,
let's try that.

This is mostly a drop-in replacement, but it does allow use to both
drop the `mysql.connector` log filtering (which prints the DB password
in plain text at log level DEBUG) and also the `playhouse.mysql_ext` stub
providing `MySQLDatabase` support for `mysql.connector` (which is where
the 9.3.0 bug we've encountered comes from).

I've done some light testing, and things seem to be working as
intented.  NOTE: the thread-local storage stuff probably isn't needed
anymore (that was necessary for `MySQLdb`, but I'll save looking in to
that for another PR.)